Fix LCD rotation and correct offset
Based on current LCD library, there is a +2,+1 offset in the mapping from the display space to the LCD.  This means some operations are clipped, and other pixels can never be written.  With CGRAM_OFFSET defined, the changes here allow the offset to be corrected.  It also fixes problems with the two "inverted" rotations which had gross errors.  These changes have been tested with a sketch writing pixels to the 4 corners (0,0), (127,127),(0,127),(127,0) in the 4 different rotations, and confirming they all display correctly using a microscope to confirm the corner pixels are the ones being activated.  This change also allows the size to be correctly set to 128,128